+++ client/src/App.js
@@ -8,6 +8,7 @@ import Login from './components/auth/Login';
 import Alert from './components/layout/Alert';
 import Dashboard from './components/dashboard/Dashboard';
 import Profiles from './components/profiles/Profiles';
+import Profile from './components/profile/Profile';
 import CreateProfile from './components/profile-forms/CreateProfile';
 import EditProfile from './components/profile-forms/EditProfile';
 import AddRestaurant from './components/profile-forms/AddRestaurant';
@@ -49,6 +50,9 @@ const App = () => {
                         </Route>
                         <Route exact path='/profiles'>
                             <Profiles />
+                        </Route>
+                        <Route exact path='/profile/:id'>
+                            <Profile />
                         </Route>
                         <PrivateRoute exact path='/dashboard'>
                             <Dashboard />